refactor: move common interface implementations to base class
