+++ picpro/protocol/IConnection.py
@@ -1,12 +1,12 @@
+import struct
 import time
 from types import TracebackType
-from typing import Optional, Type, Union, TYPE_CHECKING
+from typing import Optional, Type, Union
 import serial
 from picpro.ChipInfoEntry import ChipInfoEntry
 from picpro.exceptions import InvalidResponseError
-
-if TYPE_CHECKING:
-    from picpro.protocol.IProgrammingInterface import IProgrammingInterface
+from picpro.protocol.IProgrammingInterface import IProgrammingInterface
+from picpro.protocol.p18a.ProgrammingInterface import ProgrammingInterface
 
 
 class IConnection:
@@ -26,6 +26,11 @@ class IConnection:
             )
         except serial.SerialException as e:
             raise ConnectionError('Unable to open serial port "{}".'.format(port)) from e
+
+        self.cmd_wait_until_chip_in_socket = 0
+        self.cmd_wait_until_chip_out_of_socket = 0
+        self.cmd_programmer_version = 0
+        self.cmd_programmer_protocol = 0
 
         self.reset()
 
@@ -133,13 +138,29 @@ class IConnection:
         self.command_end()
         return result
 
+    @property
+    def expected_programmer_protocol(self) -> bytes:
+        raise NotImplementedError
+
+    def close(self) -> None:
+        self.serial_connection.close()
+
     def wait_until_chip_in_socket(self) -> None:
         """Blocks until a chip is inserted in the programming socket."""
-        raise NotImplementedError
+        cmd = 18
+        self.command_start(cmd)
+        self.expect(b'A')
+
+        self.expect(b'Y', timeout=None, send_command_end=True)
 
     def wait_until_chip_out_of_socket(self) -> None:
         """Blocks until chip is removed from programming socket."""
-        raise NotImplementedError
+        cmd = 19
+
+        self.command_start(cmd)
+        self.expect(b'A')
+
+        self.expect(b'Y', timeout=None, send_command_end=True)
 
     def programmer_version(self) -> bytes:
         """Returns the PIC programmer's numeric version.
@@ -148,21 +169,25 @@ class IConnection:
         K149-B   = 2  (Byte)
         K150     = 3  (Byte)
         """
-        raise NotImplementedError
+        cmd = 20
+        self.command_start(cmd)
+        response = self.read(1)
+        self.command_end()
+        result, = struct.unpack('B', response)
+        return result
 
     def programmer_protocol(self) -> bytes:
         """Returns the PIC programmer's protocol version in text form."""
-        raise NotImplementedError
+        cmd = 21
+        self.command_start(cmd)
+        # Protocol doc isn't clear on the format of command 22's output.
+        # Presumably it will always be exactly 4 bytes.
+        response = self.read(4)
+        self.command_end()
+        return response
 
-    @property
-    def expected_programmer_protocol(self) -> bytes:
-        raise NotImplementedError
-
-    def get_programming_interface(self, chip_info: ChipInfoEntry, icsp_mode: bool = False) -> 'IProgrammingInterface':
-        raise NotImplementedError
-
-    def close(self) -> None:
-        self.serial_connection.close()
+    def get_programming_interface(self, chip_info: ChipInfoEntry, icsp_mode: bool = False) -> IProgrammingInterface:
+        return ProgrammingInterface(self, chip_info, icsp_mode)
 
     def __enter__(self) -> 'IConnection':
         return self
